Add 'use-debounce' package to project
Added the 'use-debounce' package to the project by updating the package.json and yarn.lock files. This package will provide a debounced function that will delay the processing of the function until after wait milliseconds have elapsed since the last time it was invoked. This is essential to improve user experience by limiting the rate at which a function can fire and preventing the app from becoming unresponsive due to heavy computations.