:sparkles: Integrated QueryClient from react-query library to manage server state
This commit introduces the QueryClient from the `@tanstack/react-query` library into the main App.tsx component. This was done to enable the utilization of custom hooks for managing server state, by providing a query client that these hooks will use to fetch, cache and update data from our server. As such, it paves the way for more effective data management in the application.